requirements.txt
pip>=9
setuptools>=26
wheel>=0.29
pandas
pytest
coverage
flake8
black
yapf
python-gitlab
twine
google-cloud-storage==1.26.0
gcsfs
tensorflow.keras
numpy
matplotlib.pyplot
os
seaborn
PIL
google.colab.files
flask==1.1.1
flask-cors
cv2
scipy
